update audio-pledge service's interface to adapters
this shows another way I'm envisioning this could work. The basic idea
is that each library's adapter provides the api surface area and
triggers the events, but they also require that a sound to be operated
upon must be passed in.

Inside the audio-pledge service, instead of calling `sound.play()`, it
calls `adapter.play(sound)`. Then the adapter knows the particulars of
how to operate upon a sound of its own creation.

By delegating to the adapter, it becomes easier to reason about the
whos whats and whens of the event triggers, since they are all sent by
an adapter which manages its own sounds.