updates sound adapters to return a "soundObject"
it could go for a better name, maybe soundPackage, but it's a simple
POJO with a handle to the result of the audio library's sound creation
function and its associated adapter.

Event callbacks on the native sounds trigger events on its
associated adapter, passing in the "soundObject", which the service
unpacks and passes to our consumers.