allow consumers to specify a mime type for particular URLs
for reasons beyond a user's control, a particular audio file may not
travel with a file extension.

This is mostly a use case for streams, for instance at NY Public Radio
we serve some streams extension-less, since file extensions are not
really enforceable specs, where as content types or mime types are.

closes #8