removes canPlayExtension in favor of canPlayMimeType
this bumps up the importance of mime types in determining playability of
a given audio resource and removes canPlayExtension as a method.

the whole idea of playability is to improve performance: let
a connection specify what it can and cannot play so that we can avoid
any extraneous options during `_findFirstPlayableSound`.

so we want to continue to allow connection authors a way to specify what
will and won't play with their libraries, but still need to ease the
burden of mime type knowledge for our service consumers.

So if a mime type can not be derived from a given file path, the connection
will try to play it anyway, but send a warning to the console leading to
documentation explaining the `fileObject` syntax.