updates stream pause logic to be more reliable
I was seeing intermittent "misses" on Firefox when pausing a live
stream. After some more research, I discovered
https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/Using_HTML5_audio_and_video#Stopping_the_download_of_media.

According to MDN removing the `src` attribute and then calling load
again is the safest way to tell the browser to stop downloading this
particular piece of media. And to reliable results in Firefox too.