gate restoring internal state behind presence of sharedAudioAccess
this is only something we want to do if it's passed in. otherwise we'll
have or own audio element. this was causing bad behavior in firefox when
trying to play a sound that had ended. it would set the current time to
the final currenttime, which would trigger the `ended` event and prevent
it from playing