adds failing test for resetting isPlaying to false on audio-ended
NativeAudio took care of it due to an IE11 bux work-around that forces
us to fire `pause` when a file finishes. not so for Howler!

this patch adds a failing test and also updates the base connection to
accept event callbacks, which makes testing easier. adding an event
listener is LIFO instead of FIFO, so our callbacks were called before
the attrs are updated.