attempt to fix out of sync loading indicators
after some debugging, setting loading to false in this event wasn't
propagating out to the service, due to it already being false from some
unknown source. couldn't find out why it's getting set to false, so I'm
just sending out a notice here.

if you look inside the set method, you'll see that calling
notifyPropertyChanges is the framework's behavior if the values are
different. the reason it's not getting called is because the value is
the same at this moment.

also not setting isLoading in the loadAudio method of the native audio
connection because it's already set by the service.