dumb and dumber hifi connections
stubbing the create method to then use the create method inside the
stubbed function made for some endless call stacks. This commit changes
up how dummy sounds are generated for the unit test.

The idea is to use newly-wired DummyConnection to stand in for
sound objects with regular behavior. When we need special behavior,
a BaseSound is generated with dummy options to override baked-in
assertions.